Fixes for SyncVecEnv render and video eval normalization sync
